fix(Profile): Twitter verification error modal moved to edit
